third_party/WebKit/LayoutTests/editing/execCommand/insert_list/insert_list_in_summary_crash.html

<!-- this test requies quirks mode. -->
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable><summary><ol><li>^abc|</li></ol></summary></div>',
    'insertOrderedList',
    '<div contenteditable><summary>^abc|</summary></div>',
    'styleFromMatchedRulesForElement() should be called after distribution'));
</script>
